Split lex.rs into lex/core.rs and lex/variable.rs
`Lexer` and `lex_variable` share very little code, so it makes more
sense to organise them as separate files.
