+++ src/lex/core.rs
@@ -1,0 +1,484 @@
+use crate::lex::{END_TAG_LEN, START_TAG_LEN};
+
+enum EndTag {
+    Variable,
+    Tag,
+    Comment,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum TokenType {
+    Text,
+    Variable,
+    Tag,
+    Comment,
+}
+
+#[derive(Debug, PartialEq, Eq)]
+pub struct Token {
+    pub token_type: TokenType,
+    pub at: (usize, usize),
+}
+
+impl Token {
+    fn text(at: (usize, usize)) -> Self {
+        Self {
+            at,
+            token_type: TokenType::Text,
+        }
+    }
+
+    fn variable(at: (usize, usize)) -> Self {
+        Self {
+            at,
+            token_type: TokenType::Variable,
+        }
+    }
+
+    fn tag(at: (usize, usize)) -> Self {
+        Self {
+            at,
+            token_type: TokenType::Tag,
+        }
+    }
+
+    fn comment(at: (usize, usize)) -> Self {
+        Self {
+            at,
+            token_type: TokenType::Comment,
+        }
+    }
+}
+
+impl<'t> Token {
+    pub fn content(&self, template: &'t str) -> &'t str {
+        let (start, len) = self.at;
+        let (start, end) = match self.token_type {
+            TokenType::Text => (start, start + len),
+            TokenType::Variable => (start + START_TAG_LEN, start + len - END_TAG_LEN),
+            TokenType::Tag => (start + START_TAG_LEN, start + len - END_TAG_LEN),
+            TokenType::Comment => (start + START_TAG_LEN, start + len - END_TAG_LEN),
+        };
+        &template[start..end]
+    }
+}
+
+pub struct Lexer<'t> {
+    template: &'t str,
+    rest: &'t str,
+    byte: usize,
+    verbatim: Option<&'t str>,
+}
+
+impl<'t> Lexer<'t> {
+    pub fn new(template: &'t str) -> Self {
+        Self {
+            template,
+            rest: template,
+            byte: 0,
+            verbatim: None,
+        }
+    }
+
+    fn lex_text(&mut self) -> Token {
+        let next_tag = self.rest.find("{%");
+        let next_variable = self.rest.find("{{");
+        let next_comment = self.rest.find("{#");
+        let next = [next_tag, next_variable, next_comment]
+            .iter()
+            .filter_map(|n| *n)
+            .min();
+        let len = match next {
+            None => {
+                let len = self.rest.len();
+                self.rest = "";
+                len
+            }
+            Some(n) => {
+                self.rest = &self.rest[n..];
+                n
+            }
+        };
+        let at = (self.byte, len);
+        self.byte += len;
+        Token::text(at)
+    }
+
+    fn lex_text_to_end(&mut self) -> Token {
+        let len = self.rest.len();
+        let at = (self.byte, len);
+        self.byte += len;
+        self.rest = "";
+        Token::text(at)
+    }
+
+    fn lex_tag(&mut self, end_tag: EndTag) -> Token {
+        let end_str = match end_tag {
+            EndTag::Variable => "}}",
+            EndTag::Tag => "%}",
+            EndTag::Comment => "#}",
+        };
+        let len = match self.rest.find(end_str) {
+            None => {
+                let len = self.rest.len();
+                let at = (self.byte, len);
+                self.byte += len;
+                self.rest = "";
+                return Token::text(at);
+            }
+            Some(n) => {
+                let len = n + end_str.len();
+                self.rest = &self.rest[len..];
+                len
+            }
+        };
+        let at = (self.byte, len);
+        self.byte += len;
+        match end_tag {
+            EndTag::Variable => Token::variable(at),
+            EndTag::Tag => Token::tag(at),
+            EndTag::Comment => Token::comment(at),
+        }
+    }
+
+    fn lex_verbatim(&mut self, verbatim: &'t str) -> Token {
+        let verbatim = verbatim.trim();
+        self.verbatim = None;
+
+        let mut rest = self.rest;
+        let mut index = 0;
+        loop {
+            let next_tag = rest.find("{%");
+            match next_tag {
+                None => return self.lex_text_to_end(),
+                Some(start_tag) => {
+                    rest = &rest[start_tag..];
+                    let close_tag = rest.find("%}");
+                    match close_tag {
+                        None => return self.lex_text_to_end(),
+                        Some(end_tag) => {
+                            let inner = &rest[2..end_tag].trim();
+                            // Check we have the right endverbatim tag
+                            if inner.len() < 3 || &inner[3..] != verbatim {
+                                rest = &rest[end_tag + 2..];
+                                index += start_tag + end_tag + 2;
+                                continue;
+                            }
+
+                            index += start_tag;
+                            if index == 0 {
+                                // Return the endverbatim tag since we have no text
+                                let tag_len = end_tag + "%}".len();
+                                let at = (self.byte, tag_len);
+                                self.byte += tag_len;
+                                self.rest = &self.rest[tag_len..];
+                                return Token::tag(at);
+                            } else {
+                                self.rest = &self.rest[index..];
+                                let at = (self.byte, index);
+                                self.byte += index;
+                                return Token::text(at);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+impl Iterator for Lexer<'_> {
+    type Item = Token;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.rest.is_empty() {
+            return None;
+        }
+        Some(match self.verbatim {
+            None => match self.rest.get(..START_TAG_LEN) {
+                Some("{{") => self.lex_tag(EndTag::Variable),
+                Some("{%") => {
+                    let tag = self.lex_tag(EndTag::Tag);
+                    if let Token {
+                        token_type: TokenType::Tag,
+                        ..
+                    } = tag
+                    {
+                        let verbatim = tag.content(self.template).trim();
+                        if verbatim == "verbatim" || verbatim.starts_with("verbatim ") {
+                            self.verbatim = Some(verbatim)
+                        }
+                    }
+                    tag
+                }
+                Some("{#") => self.lex_tag(EndTag::Comment),
+                _ => self.lex_text(),
+            },
+            Some(verbatim) => self.lex_verbatim(verbatim),
+        })
+    }
+}
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn contents(template: &str, tokens: Vec<Token>) -> Vec<&str> {
+        tokens.iter().map(|t| t.content(template)).collect()
+    }
+
+    #[test]
+    fn test_lex_empty() {
+        let template = "";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![]);
+    }
+
+    #[test]
+    fn test_lex_text() {
+        let template = "Just some text";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::text((0, 14))]);
+        assert_eq!(contents(template, tokens), vec![template]);
+    }
+
+    #[test]
+    fn test_lex_text_whitespace() {
+        let template = "    ";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::text((0, 4))]);
+        assert_eq!(contents(template, tokens), vec![template]);
+    }
+
+    #[test]
+    fn test_lex_comment() {
+        let template = "{# comment #}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::comment((0, 13))]);
+        assert_eq!(contents(template, tokens), vec![" comment "]);
+    }
+
+    #[test]
+    fn test_lex_variable() {
+        let template = "{{ foo.bar|title }}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::variable((0, 19))]);
+        assert_eq!(contents(template, tokens), vec![" foo.bar|title "]);
+    }
+
+    #[test]
+    fn test_lex_tag() {
+        let template = "{% for foo in bar %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::tag((0, 20))]);
+        assert_eq!(contents(template, tokens), vec![" for foo in bar "]);
+    }
+
+    #[test]
+    fn test_lex_incomplete_comment() {
+        let template = "{# comment #";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::text((0, 12))]);
+        assert_eq!(contents(template, tokens), vec![template]);
+    }
+
+    #[test]
+    fn test_lex_incomplete_variable() {
+        let template = "{{ foo.bar|title }";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::text((0, 18))]);
+        assert_eq!(contents(template, tokens), vec![template]);
+    }
+
+    #[test]
+    fn test_lex_incomplete_tag() {
+        let template = "{% for foo in bar %";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::text((0, 19))]);
+        assert_eq!(contents(template, tokens), vec![template]);
+    }
+
+    #[test]
+    fn test_django_example() {
+        let template = "text\n{% if test %}{{ varvalue }}{% endif %}{#comment {{not a var}} {%not a block%} #}end text";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::text((0, 5)),
+                Token::tag((5, 13)),
+                Token::variable((18, 14)),
+                Token::tag((32, 11)),
+                Token::comment((43, 42)),
+                Token::text((85, 8)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![
+                "text\n",
+                " if test ",
+                " varvalue ",
+                " endif ",
+                "comment {{not a var}} {%not a block%} ",
+                "end text",
+            ]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_with_variable() {
+        let template = "{% verbatim %}{{bare   }}{% endverbatim %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::tag((0, 14)),
+                Token::text((14, 11)),
+                Token::tag((25, 17)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![" verbatim ", "{{bare   }}", " endverbatim "]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_with_tag() {
+        let template = "{% verbatim %}{% endif %}{% endverbatim %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::tag((0, 14)),
+                Token::text((14, 11)),
+                Token::tag((25, 17)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![" verbatim ", "{% endif %}", " endverbatim "]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_with_verbatim_tag() {
+        let template = "{% verbatim %}It's the {% verbatim %} tag{% endverbatim %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::tag((0, 14)),
+                Token::text((14, 27)),
+                Token::tag((41, 17)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![" verbatim ", "It's the {% verbatim %} tag", " endverbatim "]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_nested() {
+        let template = "{% verbatim %}{% verbatim %}{% endverbatim %}{% endverbatim %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::tag((0, 14)),
+                Token::text((14, 14)),
+                Token::tag((28, 17)),
+                Token::tag((45, 17)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![
+                " verbatim ",
+                "{% verbatim %}",
+                " endverbatim ",
+                " endverbatim ",
+            ]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_adjacent() {
+        let template = "{% verbatim %}{% endverbatim %}{% verbatim %}{% endverbatim %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::tag((0, 14)),
+                Token::tag((14, 17)),
+                Token::tag((31, 14)),
+                Token::tag((45, 17)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![" verbatim ", " endverbatim ", " verbatim ", " endverbatim "]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_special() {
+        let template =
+            "{% verbatim special %}Don't {% endverbatim %} just yet{% endverbatim special %}";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(
+            tokens,
+            vec![
+                Token::tag((0, 22)),
+                Token::text((22, 32)),
+                Token::tag((54, 25)),
+            ]
+        );
+        assert_eq!(
+            contents(template, tokens),
+            vec![
+                " verbatim special ",
+                "Don't {% endverbatim %} just yet",
+                " endverbatim special ",
+            ]
+        );
+    }
+
+    #[test]
+    fn test_verbatim_open_tag() {
+        let template = "{% verbatim %}Don't {% ";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::tag((0, 14)), Token::text((14, 9))]);
+        assert_eq!(contents(template, tokens), vec![" verbatim ", "Don't {% "]);
+    }
+
+    #[test]
+    fn test_verbatim_no_tag() {
+        let template = "{% verbatim %}Don't end verbatim";
+        let lexer = Lexer::new(template);
+        let tokens: Vec<_> = lexer.collect();
+        assert_eq!(tokens, vec![Token::tag((0, 14)), Token::text((14, 18))]);
+        assert_eq!(
+            contents(template, tokens),
+            vec![" verbatim ", "Don't end verbatim"]
+        );
+    }
+}
